refactor: common_metrics.py is split into metrics.py and validation.py and moved to evaluation dir and imports are updated respectively
