+++ training/model_training.py
@@ -12,7 +12,8 @@ import numpy as np
 import albumentations as A
 from tqdm import tqdm
 import gc
-from models.common_metrics import validate_all, record_validation_metrics_to_csv, getLossFunction
+from evaluation.validate import validate_all, record_validation_metrics_to_csv
+from training.loss import getLossFunction
 from models.models_tcloud import save_model_and_log_params, dict_to_hash_key,init_model_and_loaders
 import torch.nn.functional as F
 import pandas as pd
